requirements.txt
streamlit
pytesseract
pillow
opencv-python-headless
numpy
spacy
pandas
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl